feat: refactor vertical text block detection into span_orientation module
