+++ mineru/backend/utils/span_block_fix.py
@@ -1,33 +1,7 @@
 # Copyright (c) Opendatalab. All rights reserved.
 from ...types import Block, ContentType, Line, Span
 from ...utils.ocr_utils import _is_overlaps_x_exceeds_threshold, _is_overlaps_y_exceeds_threshold
-
-VERTICAL_SPAN_HEIGHT_TO_WIDTH_RATIO_THRESHOLD = 2
-VERTICAL_SPAN_IN_BLOCK_THRESHOLD = 0.8
-
-
-def is_vertical_text_block_by_spans(spans: list[Span]) -> bool:
-    """根据块内文本 span 的高宽比判断文本块是否更像竖排文本。"""
-    valid_span_count = 0
-    vertical_span_count = 0
-    for span in spans:
-        bbox = span.bbox
-        if not bbox or len(bbox) < 4:
-            continue
-
-        span_width = bbox[2] - bbox[0]
-        span_height = bbox[3] - bbox[1]
-        if span_width <= 0 or span_height <= 0:
-            continue
-
-        valid_span_count += 1
-        if span_height / span_width > VERTICAL_SPAN_HEIGHT_TO_WIDTH_RATIO_THRESHOLD:
-            vertical_span_count += 1
-
-    if valid_span_count == 0:
-        return False
-
-    return vertical_span_count / valid_span_count > VERTICAL_SPAN_IN_BLOCK_THRESHOLD
+from .span_orientation import is_vertical_text_block_by_spans
 
 
 def fix_text_block(block: Block) -> Block:
